dom/base/test/test_mozMatchesSelector.html

<!doctype html>
<meta charset=utf-8>
<title>Test for legacy mozMatchesSelector</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<div id=test></div>
<script>
test(function() {
  var element = document.getElementById("test");
  assert_true(element.matches("#test"), "matches");
  assert_true(element.mozMatchesSelector("#test"), "mozMatchesSelector");
});
</script>
